+++ lib/i18n/strings.g.dart
@@ -4,9 +4,9 @@
 /// To regenerate, run: `dart run slang`
 ///
 /// Locales: 2
-/// Strings: 2140 (1070 per locale)
+/// Strings: 2158 (1079 per locale)
 ///
-/// Built on 2026-08-18 at 02:49 UTC
+/// Built on 2026-08-18 at 12:57 UTC
 
 // coverage:ignore-file
 // ignore_for_file: type=lint, unused_import